refactor: improve item parser robustness and container logic
Hardened item definition parsing with overflow protection and safety checks. Improved container weight calculation and added developer documentation.

Changes:
- source/brushes/* — added validation for missing item definitions during brush loading
- source/game/complexitem.cpp — implemented recursive weight calculation for containers
- source/game/item.cpp — updated height logic to use sprite draw height
- source/io/otbm/item_serialization_otbm.cpp — added validation for unknown items in OTBM v1
- source/item_definitions/formats/otb/otb_item_parser.cpp — refactored attribute reading with payload helpers
- source/item_definitions/formats/dat/dat_metadata_decoder.cpp — added overflow checks for sprite counts
- source/ui/properties/* — centralized version checks for classification tiers
- source/ui/* — improved UI resilience with definition null checks

+++ source/brushes/raw/raw_brush.cpp
@@ -78,7 +78,7 @@ void RAWBrush::draw(BaseMap* map, Tile* tile, void* parameter) {
 
 	bool b = parameter ? *reinterpret_cast<bool*>(parameter) : false;
 	if ((g_settings.getInteger(Config::RAW_LIKE_SIMONE) && !b) && definition.hasFlag(ItemFlag::AlwaysOnBottom) && definition.attribute(ItemAttributeKey::AlwaysOnTopOrder) == 2) {
-		std::erase_if(tile->items, [topOrder = definition.attribute(ItemAttributeKey::AlwaysOnTopOrder)](const auto& item) {
+		std::erase_if(tile->items, [topOrder = static_cast<int>(definition.attribute(ItemAttributeKey::AlwaysOnTopOrder))](const auto& item) {
 			return item->getTopOrder() == topOrder;
 		});
 	}
